% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_binary.R
\name{arith}
\alias{arith}
\alias{+.uint8}
\alias{-.uint8}
\alias{*.uint8}
\alias{/.uint8}
\alias{\%\%.uint8}
\alias{\%/\%.uint8}
\title{Arithmetic}
\usage{
\method{+}{uint8}(e1, e2)

\method{-}{uint8}(e1, e2)

\method{*}{uint8}(e1, e2)

\method{/}{uint8}(e1, e2)

\method{\%\%}{uint8}(e1, e2)

\method{\%/\%}{uint8}(e1, e2)
}
\arguments{
\item{e1}{an atomic vector of class 'uint8'}

\item{e2}{an atomic vector of class 'uint8'}
}
\description{
Arithmetic
}
