% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{any.uint8}
\alias{all.uint8}
\alias{abs.uint8}
\alias{sign.uint8}
\alias{sum.uint8}
\alias{as.bitstring}
\alias{[.uint8}
\title{Misc ops}
\usage{
\method{any}{uint8}(e1, ...)

\method{all}{uint8}(e1, ...)

\method{abs}{uint8}(e1)

\method{sign}{uint8}(e1)

\method{sum}{uint8}(e1, ...)

as.bitstring(e1)

\method{[}{uint8}(e1, ...)
}
\description{
Misc ops
}
