% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_uint8.R
\name{order}
\alias{order}
\alias{order}
\alias{order.default}
\alias{order.uint8}
\title{Ordering permutation}
\usage{
order(...)

\method{order}{default}(...)

\method{order}{uint8}(x, ...)
}
\arguments{
\item{...}{other arguments passed to builtin 'order()'}

\item{x}{uint8 atomic vector}
}
\value{
integer atomic vector
}
\description{
Ordering permutation
}
