% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint8.R
\name{print.uint8}
\alias{print.uint8}
\title{Print a \code{uint8} vector in a friendly integer manner}
\usage{
\method{print}{uint8}(x, ...)
}
\arguments{
\item{x}{\code{uint8} vector}

\item{...}{for compatibility with default print method.}
}
\value{
Invisibly return original vector.
}
\description{
Print a \code{uint8} vector in a friendly integer manner
}
