% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_uint8.R
\name{sequence}
\alias{sequence}
\alias{:}
\alias{:.default}
\alias{:.uint8}
\title{Sequence}
\usage{
":"(from, to)

\method{:}{default}(from, to)

\method{:}{uint8}(from, to)
}
\arguments{
\item{from}{from}

\item{to}{to}
}
\value{
sequence
}
\description{
Only dispatches on first argument.
}
