% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint8.R
\name{uint8}
\alias{uint8}
\title{Create a uint8 vector of the given size}
\usage{
uint8(length = 0)
}
\arguments{
\item{length}{A non-negative integer specifying the desired length.
Double values will be coerced to integer: supplying an argument
of length other than one is an error.}
}
\value{
A \code{uint8} vector of the specified length
}
\description{
Create a uint8 vector of the given size
}
