% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{any.uint8}
\alias{all.uint8}
\alias{abs.uint8}
\alias{sign.uint8}
\alias{sum.uint8}
\alias{as.bitstring}
\alias{[.uint8}
\title{Misc ops}
\usage{
\method{any}{uint8}(..., na.rm = FALSE)

\method{all}{uint8}(..., na.rm = FALSE)

\method{abs}{uint8}(x)

\method{sign}{uint8}(x)

\method{sum}{uint8}(..., na.rm = FALSE)

as.bitstring(e1)

\method{[}{uint8}(x, ...)
}
\arguments{
\item{...}{arguments passed to other methods}

\item{na.rm}{logical. Should missing values be removed?}

\item{x}{an atomic vector of class 'uint8'}

\item{e1}{an atomic vector of class 'uint8'}
}
\description{
Misc ops
}
