% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint8.R
\name{uint8}
\alias{uint8}
\title{Create a uint8 vector}
\usage{
uint8(x)
}
\arguments{
\item{x}{Numeric vector. \code{x} will be converted to integer, and then modulo 256
to bring it in the range 0-255. NA, NaN and Inf values will cause an
error.}
}
\value{
Vector of \code{uint8} values.
}
\description{
The input is first converted to integer and this result (modulu 256) is converted
to R's built-in \code{raw} datatype.
}
